[kubernetes/apiserver] Move GetPodMetadataNames func
